+++ tol-portal-ui/app/src/pages/TUMSteps.tsx
@@ -389,7 +389,7 @@ function TUMSteps() {
           ],
         },
       }}
-      actions={['Insert into Benchling Tissue Work List', 'Mark as Not Valid']}
+      actions={['Insert into Benchling Tissue Work List', 'Mark as Not Valid', 'Export into Benchling']}
       rowSelection={true}
       {...tissueBenchling}
     />
@@ -465,7 +465,25 @@ function TUMSteps() {
           ],
         },
       }}
-      actions={['Export into Benchling', 'Mark as Not Valid']}
+      actions={[
+        'Export into Benchling',
+        'Mark as Not Valid',
+        'LRES - DNA extraction',
+        'LRES - DNA extraction (NEW)',
+        'LRES - DNA extraction + HiC',
+        'LRES - DNA extraction + HiC (NEW)',
+        'HiC',
+        'HiC (NEW)',
+        'HiC only',
+        'HiC only (NEW)',
+        'HiC + RNA',
+        'HiC + RNA (NEW)',
+        'RNA',
+        'PiMmS',
+        'RNA (NEW)',
+        'PiMmS (NEW)',
+        'Export into Benchling'
+      ]}
       rowSelection={true}
       {...sampleSTS}
     />